" Changed login and registration menus and made all ints to longs"

+++ BankApp/banking_app/src/main/java/revature/banking_app/Logic/UserVerification.java
@@ -17,7 +17,7 @@ public class UserVerification {
 
     }
 
-    public boolean verify(String username,int password){
+    public boolean verify(String username,Long password){
 
         // if the password as a string  in the database matches the password that you gave
         // as a string then it's true;
